requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.0
exceptiongroup==1.2.1
fastapi==0.110.2
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-pubsub==2.12.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.12.7
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tomli==2.0.1
trio==0.25.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
